Make pointer casts LP64-safe (drive `loses precision` errors to zero)
The decomp assumes 32-bit pointers (ILP32). Under the native LP64 build
(x86-64 / arm64), `(u32)`/`(s32)`/`(int)` casts of pointer expressions
truncate the pointer and are hard errors. This widens those casts
architecture-independently:

- Pointer arithmetic / comparison / alignment-mask / hashing casts changed
  to `(uintptr_t)` (or `(intptr_t)` where used signed). Where the result is
  a legitimately-32-bit offset/size/hash stored back into a u32 field, an
  outer narrowing cast is kept: `(u32)((uintptr_t)a - (uintptr_t)b)`.
- Round-trip pointer storage within a single header/.cpp cluster widened at
  the type level and callers adapted:
  * J3DPacket::set/getUserArea: u32 -> uintptr_t (+ PacketUtil callers)
  * JKRExpHeap allocBack `start` local: u32 -> uintptr_t
  * JKRThread / joinTwoBlocks address locals: u32 -> uintptr_t
  * JASSeqParser local `int registers[4]` -> `intptr_t`
- Shared address-conversion macros made lossless:
  OSPhysicalToCached/Uncached + OSCachedToPhysical/UncachedToPhysical
  (dolphin/os.h) and JSUConvertOffsetToPtr (JSupport.hpp).

128 eligible cast sites fixed across 47 files; full re-scan confirms zero
`loses precision` errors remain outside the event/SPC-script VM files
(EventWatcher.cpp / NpcEvent.cpp / Strategic/*), which are handled
separately. Cases where a faithful fix would require widening a HW/ARAM
address or callback parameter that ripples across many TUs were left
semantics-preserving via `(u32)(uintptr_t)ptr` and recorded for a later
coordinated widening.

+++ include/Camera/Camera.hpp
@@ -313,7 +313,7 @@ public:
 		int getThing() const
 		{
 			if (unk4 <= 0)
-				return (int)unk8; // what
+				return (int)(intptr_t)unk8; // what
 
 			return unk8[unk4 - 1];
 		}
